pyproject.toml
[project]
name = "airo-plyreader"
version = "0.2.0"
description = "GPU-accelerated point cloud pillar detection pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "open3d>=0.19.0",
    "numpy>=1.24.4",
    "cupy-cuda12x>=13.0.0",
    "cuml-cu12>=25.02",
    "matplotlib>=3.7.0",
]

[[tool.uv.index]]
url = "https://pypi.nvidia.com"
name = "nvidia"
